dashboard/requirements.txt
streamlit
python-frontmatter
pandas
requests
python-dotenv
pyyaml
